+++ modules/template/emailTemplate/commonTemplate.js
