+++ modules/template/emailTemplate/containerView.js
