+++ modules/template/emailTemplate/main-template.js
