+++ modules/template/emailTemplate/roundImage.js
