+++ modules/template/emailTemplate/updates.js
