+++ modules/template/forgotPassword.js
