+++ modules/template/redirectionFromMail.js
