+++ modules/template/sendEmailInvitation.js
