+++ modules/template/sendEmailNotification.js
