+++ modules/template/sendEmailVerification.js
